refactor: move assembleTransactionsFromMemPool
This is a utility function to fetch transactions from the mempool. Given
this has no dependency on the PeerNetwork, moving it out of the
PeerNetwork lets us clean up the class a little bit.
