+++ ironfish/src/utils/index.ts
@@ -15,8 +15,9 @@ export * from './file'
 export * from './graffiti'
 export * from './hash'
 export * from './json'
-export * from './math'
 export * from './map'
+export * from './math'
+export * from './mempool'
 export * from './node'
 export * from './promise'
 export * from './string'
